feat:major changes

+++ routes/web.php
@@ -1,10 +1,12 @@
 <?php
 
+use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ProductosController;
 use App\Http\Controllers\CategoriasController;
 use App\Http\Controllers\AlmacenesController;
 use App\Http\Controllers\MainController;
+use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 
 
@@ -18,32 +20,34 @@ use Illuminate\Http\Request;
 | be assigned to the "web" middleware group. Make something great!
 |
 */
-
 Route::get('/', [MainController::class, 'index']);
 
-Route::get('/productos', [ProductosController::class, 'index']);
-Route::post('/productos', [ProductosController::class, 'filterOptions']);
-Route::post('/filtrar/productos/{filtro}', [ProductosController::class, 'show']);
-Route::get('/crear-producto', [ProductosController::class, 'create']);
-Route::post('/crear-producto', [ProductosController::class, 'store']);
-Route::get('productos/{id}/editar', [ProductosController::class, 'edit']);
-Route::put('productos/{id}/editar', [ProductosController::class, 'update']);
-Route::delete('productos/{id}', [ProductosController::class, 'destroy']);
+Route::get('/productos', [ProductosController::class, 'index'])->middleware(['web', 'auth']);;
+Route::post('/productos', [ProductosController::class, 'filterOptions'])->middleware('auth');;
+Route::post('/filtrar/productos/{filtro}', [ProductosController::class, 'show'])->middleware('auth');;
+Route::get('/crear-producto', [ProductosController::class, 'create'])->middleware('auth');;
+Route::post('/crear-producto', [ProductosController::class, 'store'])->middleware('auth');;
+Route::get('productos/{id}/editar', [ProductosController::class, 'edit'])->middleware('auth');;
+Route::put('productos/{id}/editar', [ProductosController::class, 'update'])->middleware('auth');;
+Route::delete('productos/{id}', [ProductosController::class, 'destroy'])->middleware('auth');;
 
-Route::get('/almacenes', [AlmacenesController::class, 'index']);
-Route::get('/crear-almacen', [AlmacenesController::class, 'create']);
-Route::post('/crear-almacen', [AlmacenesController::class, 'store']);
-Route::get('almacenes/{id}/editar', [AlmacenesController::class, 'edit']);
-Route::put('almacenes/{id}/editar', [AlmacenesController::class, 'update']);
-Route::delete('almacenes/{id}', [AlmacenesController::class, 'destroy']);
+Route::get('/almacenes', [AlmacenesController::class, 'index'])->middleware('auth');;
+Route::get('/crear-almacen', [AlmacenesController::class, 'create'])->middleware('auth');;
+Route::post('/crear-almacen', [AlmacenesController::class, 'store'])->middleware('auth');;
+Route::get('almacenes/{id}/editar', [AlmacenesController::class, 'edit'])->middleware('auth');;
+Route::put('almacenes/{id}/editar', [AlmacenesController::class, 'update'])->middleware('auth');;
+Route::delete('almacenes/{id}', [AlmacenesController::class, 'destroy'])->middleware('auth');;
 
-Route::get('/categorias', [CategoriasController::class, 'index']);
-Route::get('/crear-categoria', [CategoriasController::class, 'create']);
-Route::post('/crear-categoria', [CategoriasController::class, 'store']);
-Route::get('categorias/{id}/editar', [CategoriasController::class, 'edit']);
-Route::put('categorias/{id}/editar', [CategoriasController::class, 'update']);
-Route::delete('categorias/{id}', [CategoriasController::class, 'destroy']);
+Route::get('/categorias', [CategoriasController::class, 'index'])->middleware('auth');;
+Route::get('/crear-categoria', [CategoriasController::class, 'create'])->middleware('auth');;
+Route::post('/crear-categoria', [CategoriasController::class, 'store'])->middleware('auth');;
+Route::get('categorias/{id}/editar', [CategoriasController::class, 'edit'])->middleware('auth');;
+Route::put('categorias/{id}/editar', [CategoriasController::class, 'update'])->middleware('auth');;
+Route::delete('categorias/{id}', [CategoriasController::class, 'destroy'])->middleware('auth');;
 
-//Route::resource('productos', App\Http\Controllers\ProductosController::class);
-
+Route::get('/login', [UserController::class, 'index'])->name('login');
+Route::post('/login', [UserController::class, 'authenticate']);
+Route::post('/logout', [UserController::class, 'logout']);
+Route::get('/signup', [UserController::class, 'create'])->name('signup');
+Route::post('/signup', [UserController::class, 'store']);
